Runs repeatedly using specified interval on cli
Can now execute 'synopsys 1' and the command will repeat output until
sigint or sigterm is received.

To do this made changes in all the temporary 'infoprint' methods to just
return a string.

Also updated CPUTime stats to track the previous timepoint and do maths
to get a fraction of each usage category. So now show a value of 0-1 for
usr, idle, sys etc..